refactor(opencode): align Result.try error form and move branded types to _types.ts
- Change Result.try catch from new Error(String(error), { cause: error }) to
  identity form (error) => error, matching codex/amp convention
- Use String(error) instead of .message for logger.debug calls
- Move modelNameSchema/sessionIdSchema from data-loader.ts to _types.ts
  with exported creator functions (createModelName, createSessionId)
  and type aliases (ModelName, SessionId), following ccusage pattern
